+++ 1. HTTP/http1_1/serverHTTP1_1.c
@@ -7,6 +7,10 @@
 #include <unistd.h>
 #include <fcntl.h>
 
+/**
+ * Funzione di utilità per garantire l'invio di tutti i byte richiesti.
+ * Gestisce eventuali invii parziali da parte della system call write().
+ */
 void sendByte(int fd, char *buffer, int numeroByte)
 {
     int m = 0;
@@ -22,6 +26,7 @@ void sendByte(int fd, char *buffer, int numeroByte)
 
 int main()
 {
+    // Struttura per memorizzare gli header della richiesta
     struct header
     {
         char *n;
@@ -30,16 +35,19 @@ int main()
 
     struct header h[100];
 
+    // Creazione del socket del server (TCP)
     int sockfd = socket(AF_INET, SOCK_STREAM, 0);
 
+    // Configurazione del socket per permettere il riutilizzo immediato dell'indirizzo/porta dopo la chiusura
     int opt = 1;
     setsockopt(sockfd, SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(opt));
 
     struct sockaddr_in address;
     address.sin_family = AF_INET;
-    address.sin_port = htons(8081);
-    address.sin_addr.s_addr = INADDR_ANY;
+    address.sin_port = htons(8081);       // Il server ascolta sulla porta 8081
+    address.sin_addr.s_addr = INADDR_ANY; // Accetta connessioni da qualsiasi interfaccia di rete
 
+    // Associazione del socket all'indirizzo e alla porta configurati
     int s = bind(sockfd, (struct sockaddr *)&address, sizeof(address));
 
     if (s != 0)
@@ -53,20 +61,24 @@ int main()
         fflush(stdout);
     }
 
+    // Messa in ascolto del socket: coda di massimo 5 connessioni pendenti
     int l = listen(sockfd, 5); // posso accettare 5 client in parallelo
 
     printf("il server è pronto per l'accept\n");
     fflush(stdout);
 
+    // Loop principale per accettare le connessioni dei client
     while (1)
     {
         int clientSockId = accept(sockfd, NULL, NULL);
 
+        // Creazione di un processo figlio per gestire la richiesta specifica
         int forkId = fork(); // creo un processo figlio per gestire la richiesta del client, così il processo padre può continuare ad accettare altre richieste
         if (forkId == 0)     // entro nel processo figlio, gestisco la richiesta del client
         {
             printf("accept effettuata\n");
 
+            // Buffer per leggere la richiesta HTTP
             // Buffer aumentato per poter leggere interamente richieste con body corposi (es. POST)
             char buffer[10000];
 
@@ -78,28 +90,33 @@ int main()
             // int byteLetti = 0;
             int headerIndex = 0;
 
+            // Parsing della richiesta byte per byte fino alla fine degli header (\r\n\r\n)
             // Ciclo per leggere la richiesta carattere per carattere fino alla fine degli header (\r\n\r\n)
             while ((n += read(clientSockId, buffer + n, 1)) > 0)
             {
+                // Fine di una riga negli header
                 if (buffer[n - 1] == '\n' && buffer[n - 2] == '\r')
                 {
+                    // Controllo se abbiamo trovato la riga vuota (\r\n\r\n) che separa header e body
                     if (buffer[n - 4] == 0)
                     {
                         // Raggiunta la riga vuota che separa header e body (\r\n\r\n)
                         break;
                     }
                     lettoNomeHeader = 0;
-                    buffer[n - 2] = 0;
-                    h[headerIndex].n = buffer + n;
+                    buffer[n - 2] = 0;             // Termino la stringa del valore dell'header precedente
+                    h[headerIndex].n = buffer + n; // Inizio potenziale del prossimo nome header
                 }
+                // Identificazione del separatore ':' tra nome e valore dell'header
                 else if (!lettoNomeHeader && buffer[n - 1] == ':')
                 {
                     lettoNomeHeader = 1;
-                    buffer[n - 1] = 0;
-                    h[headerIndex++].v = buffer + n;
+                    buffer[n - 1] = 0;               // Termino la stringa del nome dell'header
+                    h[headerIndex++].v = buffer + n; // Inizio del valore dell'header
                 }
             }
 
+            // Ricerca del Content-Length per gestire eventuali body (es. in POST)
             // Variabile per salvare la dimensione del body (fondamentale per le richieste POST)
             int contentLength = 0;
             printf("stampo gli headers\n");
@@ -114,6 +131,7 @@ int main()
             }
             printf("content length = %d\n", contentLength);
 
+            // Analisi della Request Line (es: "GET /index.html HTTP/1.1")
             char *requestLine = buffer; // Es. estrazione di: GET /index.html HTTP/1.1
             printf("request line: %s\n", requestLine);
 
@@ -127,54 +145,65 @@ int main()
 
             char response[1024];
 
+            // --- GESTIONE METODO GET ---
             if (strcmp(method, "GET") == 0) // Gestione del metodo GET
             {
+                // Utilizziamo Transfer-Encoding: chunked per inviare i file
                 // Prepara la risposta con Transfer-Encoding: chunked per inviare il file a blocchi
                 strcpy(response, "HTTP/1.1 200 OK\r\nTransfer-Encoding: chunked\r\n\r\n");
 
+                // Default sulla home page
                 if (strcmp(uri, "/") == 0)
                     sprintf(uri, "/index.html");
 
+                // Tentativo di apertura del file richiesto
                 int fd = open(uri + 1, O_RDONLY); // uri + 1 per saltare la barra iniziale (es. "/index.html" -> "index.html")
 
                 if (fd < 0)
                 {
+                    // File non trovato: errore 404
                     // Se il file non esiste, preparo e invio l'errore 404
                     sprintf(response, "HTTP/1.1 404 Not Found\r\n\r\n<html><body>Pagina non trovata!</body></html>");
-                    sendByte(clientSockId, response, strlen(response));
+                    sendByte(clientSockId, response, (int)strlen(response));
                 }
                 else
                 {
+                    // Invio degli header iniziali
                     // Chiamo la funzione definita sopra per inviare gli header HTTP al client
-                    sendByte(clientSockId, response, strlen(response));
+                    sendByte(clientSockId, response, (int)strlen(response));
 
                     int m = 0;
                     char bufferFile[1024];
 
+                    // Lettura del file e invio a blocchi (chunks)
                     // Leggo il file e lo invio a blocchi (chunk) secondo lo standard HTTP/1.1
                     while ((m = read(fd, bufferFile, sizeof(bufferFile))) > 0)
                     {
                         char chunkSize[100];
+                        // La dimensione del chunk deve essere scritta in esadecimale
                         sprintf(chunkSize, "%x\r\n", m); // Scrivo la dimensione del chunk in esadecimale seguita da \r\n
                         printf("sto inviando m = %d, 0x%s byte ", m, chunkSize);
 
-                        sendByte(clientSockId, chunkSize, strlen(chunkSize)); // Invio dimensione chunk
-                        sendByte(clientSockId, bufferFile, m);                // Invio contenuto chunk
-                        sendByte(clientSockId, "\r\n", 2);                    // Invio terminatore chunk
+                        sendByte(clientSockId, chunkSize, (int)strlen(chunkSize)); // Invio dimensione chunk
+                        sendByte(clientSockId, bufferFile, m);                     // Invio contenuto chunk
+                        sendByte(clientSockId, "\r\n", 2);                         // Ogni chunk termina con CRLF
                     }
+                    // Invio del chunk finale di dimensione 0 per terminare la risposta chunked
                     // Invio il chunk vuoto di dimensione 0 per indicare al client la fine della trasmissione
                     sendByte(clientSockId, "0\r\n\r\n", 5);
                 }
 
                 close(fd);
             }
+            // --- GESTIONE METODO POST ---
             else if (strcmp(method, "POST") == 0) // Gestione del metodo POST
             {
                 int m = 0;
                 char bufferFile[10000] = {0};
 
+                // Lettura del corpo della richiesta basata sulla Content-Length ricevuta
                 // Leggo il body della richiesta iterando finché non ho letto tutti i byte indicati dal Content-Length
-                while (m < contentLength && m < sizeof(bufferFile) - 1)
+                while (m < contentLength && m < (int)sizeof(bufferFile) - 1)
                 {
                     int r = read(clientSockId, bufferFile + m, sizeof(bufferFile) - 1 - m);
                     if (r <= 0)
@@ -184,22 +213,26 @@ int main()
 
                 printf("buffer Body:\n%s\n", bufferFile);
 
+                // Risposta di conferma ricezione dati POST
                 sprintf(response, "HTTP/1.1 200 OK\r\n\r\n<html><body>Dati POST ricevuti correttamente!</body></html>");
                 // Chiamo la funzione definita sopra per inviare la risposta al client
-                sendByte(clientSockId, response, strlen(response));
+                sendByte(clientSockId, response, (int)strlen(response));
             }
+            // --- METODO NON SUPPORTATO ---
             else // Se il metodo non è né GET né POST (es. DELETE, PUT)
             {
                 sprintf(response, "HTTP/1.1 405 Method Not Allowed\r\n\r\n");
                 // Chiamo la funzione definita sopra per inviare la risposta di errore al client
-                sendByte(clientSockId, response, strlen(response));
+                sendByte(clientSockId, response, (int)strlen(response));
             }
 
+            // Chiusura del socket del client nel processo figlio e uscita
             close(clientSockId);
             return 0; // Termina il processo figlio
         }
         else
         {
+            // Il processo padre chiude il socket del client e torna in ascolto (accept)
             close(clientSockId); // Nel processo padre, chiudo il socket dedicato al client e mi rimetto in attesa sulla listen
         }
     }
